[FIX] Encoding charset default is not always UTF-8 (#472).
